+++ public/src/components/channelManagement/epicTests/epicTestVariantEditor.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import {EpicVariant} from './epicTestsForm';
 import {Cta, defaultCta} from '../helpers/shared';
 import {Theme, createStyles, WithStyles, withStyles} from "@material-ui/core";
-import EditableTextField from "../helpers/editableTextField";
+import EditableTextField from "../editableTextField";
 import CtaEditor from "../ctaEditor";
 import TickerEditor from '../tickerEditor';
-import ButtonWithConfirmationPopup from '../helpers/buttonWithConfirmationPopup';
+import ButtonWithConfirmationPopup from '../buttonWithConfirmationPopup';
 import DeleteSweepIcon from '@material-ui/icons/DeleteSweep';
 import {onFieldValidationChange, ValidationStatus} from '../helpers/validation';
 import {getInvalidTemplateError} from '../helpers/copyTemplates';
